refactor: Comment 단위 테스트 경로변경
