+++ src/test/java/com/sprint5team/monew/repository/comment/CommentRepositoryTest.java
@@ -1,7 +1,8 @@
-package com.sprint5team.monew.domain.comment.repository;
+package com.sprint5team.monew.repository.comment;
 
 import com.sprint5team.monew.domain.article.entity.Article;
 import com.sprint5team.monew.domain.comment.entity.Comment;
+import com.sprint5team.monew.domain.comment.repository.CommentRepository;
 import com.sprint5team.monew.domain.user.entity.User;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
